+++ packages/api-contract/src/promise/Code.ts
@@ -4,13 +4,13 @@
 import type { ApiPromise } from '@polkadot/api';
 import type { AnyJson } from '@polkadot/types/types';
 
-import { decorateMethodPromise } from '@polkadot/api';
+import { toPromiseMethod } from '@polkadot/api';
 
 import { Abi } from '../Abi';
 import { Code as BaseCode } from '../base';
 
 export class Code extends BaseCode<'promise'> {
   constructor (api: ApiPromise, abi: AnyJson | Abi, wasm: Uint8Array | string | Buffer | null | undefined) {
-    super(api, abi, wasm, decorateMethodPromise);
+    super(api, abi, wasm, toPromiseMethod);
   }
 }