+++ packages/api-contract/src/promise/Contract.ts
@@ -4,13 +4,13 @@
 import type { AccountId } from '@polkadot/types/interfaces';
 import type { AnyJson } from '@polkadot/types/types';
 
-import { ApiPromise, decorateMethodPromise } from '@polkadot/api';
+import { ApiPromise, toPromiseMethod } from '@polkadot/api';
 
 import { Abi } from '../Abi';
 import { Contract as BaseContract } from '../base';
 
 export class Contract extends BaseContract<'promise'> {
   constructor (api: ApiPromise, abi: AnyJson | Abi, address: string | AccountId) {
-    super(api, abi, address, decorateMethodPromise);
+    super(api, abi, address, toPromiseMethod);
   }
 }